Day 5: part 1 solved
I wasn't using tail recursion, so blew the stack. I tried to find a solution
chunking int pairs or working pairwise through the list, but the way
a reaction effects the rest of the sequence was giving me problems.

Found a `fold` solution that I like a lot that's really fast